+++ trunk/tahoe/src/elements/continuum/solid/materials/Simo-Holzapfel/KStV/SSSV_KStV2D.cpp
@@ -1,4 +1,4 @@
-/* $Id: SSSV_KStV2D.cpp,v 1.2 2002-10-05 03:44:33 paklein Exp $ */
+/* $Id: SSSV_KStV2D.cpp,v 1.3 2002-10-14 16:04:07 thao Exp $ */
 /* created:   TDN (5/31/2001) */
 
 #include "SSSV_KStV2D.h"
@@ -28,8 +28,8 @@ SSSV_KStV2D::SSSV_KStV2D(ifstreamT& in, const SmallStrainT& element):
 	in >> ftauS;
 	in >> ftauB;
 
-	int taudtS = fdt/ftauS;
-	int taudtB = fdt/ftauB;
+	double taudtS = fdt/ftauS;
+	double taudtB = fdt/ftauB;
 
 	falphaS = exp(-0.5*taudtS);
 	falphaB = exp(-0.5*taudtB);
